requirements.txt
streamlit>=1.32.0
torch>=2.0
transformers>=4.30
openai-whisper
Pillow>=9.0
pyttsx3
scikit-learn
numpy
scipy
